prep: Copy helpers and paste them to metadata
To prepare refactoring.
No modifications at all.
It doesn't even compile.
